// src/types/subscribed-form-register.ts
export type RegisterSubscribedFormType = {
    fullName: string
    badgeName: string
    password: string
    confirmPassword: string
    cpf: string
    birthDate: string
    cep: string
    address: string
    number: string
    neighborhood: string
    city: string
    state: string
    email: string
    phone: string
    mobile: string
}